+++ packages/eslint-config-ali/test/fixtures/ts-import-a.ts
@@ -1,0 +1,3 @@
+import { b } from '@/ts-import-b';
+
+export const a = b;